+++ src/main/java/net/lugo/utools/features/Zoom.java
@@ -4,15 +4,15 @@ import net.lugo.utools.UTools;
 import net.minecraft.client.MinecraftClient;
 
 public class Zoom {
-    private static final double MIN_ZOOM = 1D;
-    private static final double MAX_ZOOM = 100D;
+    private static final float MIN_ZOOM = 1f;
+    private static final float MAX_ZOOM = 100f;
     
     private static final MinecraftClient MC = MinecraftClient.getInstance();
     
     public static boolean isZooming = false;
-    public static double goal = 1D;
-    public static double lastGoal = 1D;
-    public static double latestEffectiveZoom = 1D;
+    public static float goal = 1f;
+    public static float lastGoal = 1f;
+    public static float latestEffectiveZoom = 1f;
     public static float t = 0f;
 
     public static void on () {
@@ -26,10 +26,10 @@ public class Zoom {
         isZooming = false;
         lastGoal = latestEffectiveZoom;
         t = 0f;
-        goal = 1D;
+        goal = 1f;
         MC.options.smoothCameraEnabled = false;
     }
-    public static void increment (Double amount) {
+    public static void increment (float amount) {
         lastGoal = latestEffectiveZoom;
         t = 0f;
         goal = Math.clamp(goal + amount, MIN_ZOOM, MAX_ZOOM);
